requirements.txt
streamlit>=1.31.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
chromadb>=0.4.22
pandas>=2.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0  # Changed from docx to python-docx
requests>=2.31.0
flask>=3.0.0